globally renamed Stylometry.py to stylometry.py, to be more in-line with the python naming scheme
